+++ src/main/java/com/example/MvcReference/controller/ViewController.java
@@ -11,13 +11,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import com.example.MvcReference.entity.Buku;
-import com.example.MvcReference.service.BukuServiceImplement;
+import com.example.MvcReference.service.implement.BukuServiceImplement;
 
 @Controller
 // controller
 public class ViewController {
     @Autowired
     private BukuServiceImplement bukuService;
+
 
     @GetMapping("/login")
     public String login(Model model) {
